requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.2.0
xgboost>=1.7.0
yfinance>=0.2.0
python-dotenv>=0.21.0
joblib>=1.2.0
